copilot/add helios and sol extensions
